requirements.txt
addict==2.4.0
albumentations==1.3.1
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
cityscapesScripts==2.2.2
click==8.1.3
cmake==3.26.4
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
ConfigArgParse==1.5.5
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
Cython==0.29.36
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flake8==5.0.4
Flask==2.2.5
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
imagecorruptions==1.1.2
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
interrogate==1.5.0
ipykernel==6.24.0
ipython==8.12.2
ipywidgets==8.0.7
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
kwarray==0.6.13
lazy_loader==0.3
lit==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
# Editable install with no version control (mmcv-full==1.5.0)
-e /home/kelin/Documents/swin/mmcv-1.5.0
# Editable install with no version control (mmdet==2.11.0)
-e /home/kelin/Documents/swin
mmengine==0.8.0
mmlvis==10.5.3
model-index==0.1.11
mpmath==1.3.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.7.0
onnxruntime==1.5.1
open3d==0.17.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
plotly==5.15.0
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycocotools==2.0.6
pycodestyle==2.9.1
pycryptodome==3.18.0
pyflakes==2.5.0
Pygments==2.15.1
pyparsing==3.1.0
pyquaternion==0.9.9
pyrsistent==0.19.3
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
requests==2.31.0
retrying==1.3.4
rich==13.4.2
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
shapely==2.0.1
shiboken6==6.5.1.1
six==1.16.0
sklearn==0.0.post5
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
ubelt==1.3.2
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
xdoctest==1.1.1
yapf==0.40.1
zipp==3.15.0
